requirements.txt
numpy>=1.26,<2.0
pandas>=2.0
scikit-learn>=1.3
scipy>=1.11
matplotlib>=3.7
seaborn>=0.12
xgboost>=2.0
tensorflow>=2.15,<2.16
keras>=2.15,<2.16
joblib>=1.3
flask>=3.0
flask-cors>=4.0
streamlit>=1.30
plotly>=5.18
pyyaml>=6.0
pillow>=10.0
psycopg2-binary>=2.9
